Remove old renderer and create new shell
