+++ shell/src/lib.rs
@@ -1,8 +1,9 @@
-use std::cell::RefCell;
-
 use anyhow::Result;
-use cosmic_text::{FontSystem, SwashCache};
-use granularity::{Producer, Value};
+use cosmic_text as text;
+use cosmic_text::FontSystem;
+use granularity_geometry::{scalar, Camera, Matrix4};
+use granularity_renderer::{Renderer, ShapeRenderer, ShapeRendererContext};
+use granularity_shapes::Shape;
 use log::{error, info};
 use wgpu::{CommandBuffer, PresentMode, SurfaceTexture};
 use winit::{
@@ -11,24 +12,99 @@ use winit::{
     window::{Window, WindowBuilder},
 };
 
-#[derive(Clone)]
-pub struct Shell {
-    pub font_system: Value<RefCell<FontSystem>>,
-    pub glyph_cache: Value<RefCell<SwashCache>>,
-    pub surface: Value<wgpu::Surface>,
-    pub device: Value<wgpu::Device>,
-    pub queue: Value<wgpu::Queue>,
-    pub surface_config: Value<wgpu::SurfaceConfiguration>,
-}
-
 pub trait Application {
     fn update(&mut self, window_event: WindowEvent<'static>);
-    fn runtime(&self) -> granularity::Runtime;
+    fn render(&self, shell: &mut Shell) -> (Camera, Vec<Shape>);
+}
+
+const Z_RANGE: (scalar, scalar) = (0.1, 100.0);
+
+pub async fn run<A: Application + 'static>(mut application: A) -> Result<()> {
+    let event_loop = EventLoop::new();
+    let window = WindowBuilder::new().build(&event_loop).unwrap();
+    let mut shell = Shell::new(&window).await;
+
+    event_loop.run(move |event, _, control_flow| {
+        match event {
+            Event::WindowEvent { event, window_id } if window_id == window.id() => match event {
+                WindowEvent::CloseRequested
+                | WindowEvent::KeyboardInput {
+                    input:
+                        KeyboardInput {
+                            state: ElementState::Pressed,
+                            virtual_keycode: Some(VirtualKeyCode::Escape),
+                            ..
+                        },
+                    ..
+                } => *control_flow = ControlFlow::Exit,
+                WindowEvent::Resized(physical_size) => {
+                    shell.resize_surface((physical_size.width, physical_size.height));
+                    window.request_redraw()
+                }
+                WindowEvent::ScaleFactorChanged { new_inner_size, .. } => {
+                    shell.resize_surface((new_inner_size.width, new_inner_size.height));
+                    window.request_redraw()
+                }
+                event => {
+                    if let Some(static_event) = event.to_static() {
+                        application.update(static_event);
+                        window.request_redraw()
+                    }
+                }
+            },
+            Event::RedrawRequested(window_id) if window_id == window.id() => {
+                let (camera, shapes) = application.render(&mut shell);
+                let surface_matrix = shell.renderer.surface_matrix();
+                let surface_size = shell.renderer.surface_size();
+                // TODO: This is a mess.
+                let mut shape_renderer_context = ShapeRendererContext {
+                    device: &shell.renderer.device,
+                    queue: &shell.renderer.queue,
+                    texture_sampler: &shell.renderer.texture_sampler,
+                    texture_bind_group_layout: &shell.renderer.texture_bind_group_layout,
+                    font_system: &mut shell.font_system,
+                };
+                let view_projection_matrix = camera.view_projection_matrix(Z_RANGE, surface_size);
+                let primitives = shell.shape_renderer.render(
+                    &mut shape_renderer_context,
+                    &view_projection_matrix,
+                    &surface_matrix,
+                    &shapes,
+                );
+                // TODO: pass primitives as value.
+                match shell
+                    .renderer
+                    .render_and_present(&view_projection_matrix, &primitives)
+                {
+                    Ok(_) => {}
+                    // Reconfigure the surface if lost
+                    // TODO: shouldn't we redraw here? Also, I think the renderer can do this, too.
+                    Err(wgpu::SurfaceError::Lost) => shell.reconfigure_surface(),
+                    // The system is out of memory, we should probably quit
+                    Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
+                    // All other errors (Outdated, Timeout) should be resolved by the next frame
+                    Err(e) => error!("{:?}", e),
+                }
+            }
+            Event::MainEventsCleared => {
+                // RedrawRequested will only trigger once, unless we manually
+                // request it.
+                // window.request_redraw();
+            }
+            _ => {}
+        }
+    });
+}
+
+pub struct Shell {
+    pub font_system: text::FontSystem,
+    shape_renderer: ShapeRenderer,
+    renderer: Renderer,
 }
 
 impl Shell {
     // Creating some of the wgpu types requires async code
-    async fn new(runtime: granularity::Runtime, window: &Window) -> Shell {
+    async fn new(window: &Window) -> Shell {
         let size = window.inner_size();
 
         // The instance is a handle to our GPU
@@ -89,7 +165,7 @@ impl Shell {
             .unwrap_or(surface_caps.present_modes[0]);
 
         info!("Selecting present mode {:?}", present_mode);
-        let config = wgpu::SurfaceConfiguration {
+        let surface_config = wgpu::SurfaceConfiguration {
             usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
             format: *surface_format,
             width: size.width,
@@ -99,69 +175,39 @@ impl Shell {
             alpha_mode: surface_caps.alpha_modes[0],
             view_formats: vec![],
         };
-        surface.configure(&device, &config);
+        surface.configure(&device, &surface_config);
 
-        let font_system = runtime.var(FontSystem::new().into());
-        let glyph_cache = runtime.var(SwashCache::new().into());
-        // TODO: analyze dependencies and integrate the shell and its updates into the graph.
-        let surface = runtime.var(surface);
-        let device = runtime.var(device);
-        let queue = runtime.var(queue);
-        let config = runtime.var(config);
+        let font_system = FontSystem::new();
+
+        let renderer = Renderer::new(device, queue, surface, surface_config);
 
         Self {
             font_system,
-            glyph_cache,
-            surface,
-            device,
-            queue,
-            surface_config: config,
+            shape_renderer: ShapeRenderer::default(),
+            renderer,
         }
     }
 
+    /// A Matrix that translates from pixels (0,0)-(width,height) to screen space, which is -1.0 to
+    /// 1.0 in each axis. Also flips y.
+    pub fn pixel_matrix(&self) -> Matrix4 {
+        let (_, surface_height) = self.renderer.surface_size();
+        Matrix4::from_nonuniform_scale(1.0, -1.0, 1.0)
+            * Matrix4::from_scale(1.0 / surface_height as f64 * 2.0)
+    }
+
     fn resize_surface(&mut self, new_size: (u32, u32)) {
-        let new_surface_size = (new_size.0.max(1), new_size.1.max(1));
-
-        if new_surface_size != self.surface_size() {
-            self.surface_config.apply(|mut config| {
-                config.width = new_surface_size.0;
-                config.height = new_surface_size.1;
-                config
-            });
-
-            self.reconfigure_surface();
-        }
+        self.renderer.resize_surface(new_size);
     }
 
     /// Reconfigure the surface after a change to the window's size or format.
     fn reconfigure_surface(&mut self) {
-        self.surface.apply(|surface| {
-            surface.configure(&self.device.get_ref(), &self.surface_config.get_ref());
-            surface
-        })
+        self.renderer.reconfigure_surface()
     }
 
     // Surface size may not match the Window's size, for example if the window's size is 0,0.
     fn surface_size(&self) -> (u32, u32) {
-        let config = self.surface_config.get_ref();
-        (config.width, config.height)
-    }
-
-    fn update(&mut self) {}
-
-    fn render(
-        &mut self,
-        command_buffer: &mut Value<CommandBuffer>,
-        surface_texture: &mut Value<SurfaceTexture>,
-    ) -> Result<(), wgpu::SurfaceError> {
-        self.queue.get_ref().submit([command_buffer.take()]);
-        surface_texture.take().present();
-
-        Ok(())
-    }
-
-    pub fn runtime(&self) -> granularity::Runtime {
-        self.font_system.runtime()
+        self.renderer.surface_size()
     }
 }
 
@@ -172,65 +218,7 @@ pub fn time<T>(name: &str, f: impl FnOnce() -> T) -> T {
     r
 }
 
-pub async fn run<A: Application + 'static>(
-    mut application: A,
-    create_render_graph: impl FnOnce(&A, &Shell) -> (Value<wgpu::CommandBuffer>, Value<wgpu::SurfaceTexture>)
-        + 'static,
-) -> Result<()> {
-    let event_loop = EventLoop::new();
-    let window = WindowBuilder::new().build(&event_loop).unwrap();
-    let runtime = application.runtime();
-    let mut shell = Shell::new(runtime, &window).await;
-
-    let (mut command_buffer, mut surface_texture) = create_render_graph(&application, &shell);
-
-    event_loop.run(move |event, _, control_flow| {
-        match event {
-            Event::WindowEvent { event, window_id } if window_id == window.id() => match event {
-                WindowEvent::CloseRequested
-                | WindowEvent::KeyboardInput {
-                    input:
-                        KeyboardInput {
-                            state: ElementState::Pressed,
-                            virtual_keycode: Some(VirtualKeyCode::Escape),
-                            ..
-                        },
-                    ..
-                } => *control_flow = ControlFlow::Exit,
-                WindowEvent::Resized(physical_size) => {
-                    shell.resize_surface((physical_size.width, physical_size.height));
-                    window.request_redraw()
-                }
-                WindowEvent::ScaleFactorChanged { new_inner_size, .. } => {
-                    shell.resize_surface((new_inner_size.width, new_inner_size.height));
-                    window.request_redraw()
-                }
-                event => {
-                    if let Some(static_event) = event.to_static() {
-                        application.update(static_event);
-                        window.request_redraw()
-                    }
-                }
-            },
-            Event::RedrawRequested(window_id) if window_id == window.id() => {
-                shell.update();
-                match shell.render(&mut command_buffer, &mut surface_texture) {
-                    Ok(_) => {}
-                    // Reconfigure the surface if lost
-                    // TODO: shouldn't we redraw here?
-                    Err(wgpu::SurfaceError::Lost) => shell.reconfigure_surface(),
-                    // The system is out of memory, we should probably quit
-                    Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
-                    // All other errors (Outdated, Timeout) should be resolved by the next frame
-                    Err(e) => error!("{:?}", e),
-                }
-            }
-            Event::MainEventsCleared => {
-                // RedrawRequested will only trigger once, unless we manually
-                // request it.
-                // window.request_redraw();
-            }
-            _ => {}
-        }
-    });
+fn main() {
+    // TODO: Need to put the shell into its own crate.
+    println!("Run the hello example instead")
 }
