--- text/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "granularity-text"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-granularity-shell = { workspace = true }
-bitflags = { workspace = true }
-wgpu = { workspace = true }
-anyhow = { workspace = true }
-granularity = { workspace = true }
-cgmath = { workspace = true }
-bytemuck = { workspace = true }
-granularity-geometry = { workspace = true }
-cosmic-text = { workspace = true }
-# TODO: remove (use only cosmic-text)
-swash = { workspace = true }
-log = { workspace = true }
-nearly = { workspace = true }
-derive_more = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true }
-env_logger = { workspace = true }
-winit = { workspace = true }
