--- text/src/lib.rs
@@ -1,13 +1,0 @@
-#![allow(clippy::needless_range_loop)]
-
-mod distance_field_gen;
-mod label;
-mod layout;
-mod point;
-mod render_graph;
-mod vertex;
-
-// pub use label::*;
-pub use layout::*;
-// pub use render_graph::render_graph;
-pub use vertex::*;
